Moved send_message method to payments from bot
