+++ ducatus_exchange/payments/models.py
@@ -1,9 +1,13 @@
-from django.db import models
+import logging
+
+from django.db import models, transaction
 from django_fsm import FSMField, transition
 
-from ducatus_exchange.bot.services import send_or_update_message
+from ducatus_exchange.bot.models import BotSub, BotSwapMessage
+# from ducatus_exchange.bot.services import send_or_update_message
 from ducatus_exchange.consts import MAX_DIGITS
 from ducatus_exchange.exchange_requests.models import ExchangeRequest
+from ducatus_exchange.start_bot_polling import Bot
 
 
 class Payment(models.Model):
@@ -31,27 +35,51 @@ class Payment(models.Model):
     collection_tx_hash = models.CharField(max_length=100, null=True, default='')
     returned_tx_hash = models.CharField(max_length=100, null=True, default='')
 
+    def generate_message(swap_id, state):
+        return f'Payment: {swap_id} handled with status: {state}'
+
+    @transaction.atomic
+    def send_or_update_message(self, swap_id, state):
+        logger = logging.getLogger('bot')    
+        subs = BotSub.objects.all()
+        message = self.generate_message(swap_id, state)
+        for sub in subs:
+            try:
+                message_model, created = BotSwapMessage.objects.select_for_update().get_or_create(
+                    swap_id=swap_id, sub=sub
+                )
+                if created:
+                    message_model.message_id = Bot.bot.send_message(sub.chat_id, message, parse_mode='html',
+                                                                        disable_web_page_preview=True).message_id
+                    message_model.save()
+                else:
+                    Bot.bot.edit_message_text(message, sub.chat_id, message_model.message_id, parse_mode='html',
+                                                    disable_web_page_preview=True)
+            except Exception as e:
+                logger.error(msg=f'send_or_update_message FAILED on payment: {swap_id} with exception: \n {e}')
+        logger.info(msg=f'send_or_update_message SUCCEDED on payment: {swap_id}')
+
     # States change
     @transition(field=transfer_state, source=['WAITING_FOR_TRANSFER', 'ERROR'], target='DONE')
     def state_transfer_done(self):
-        send_or_update_message(self.tx_hash, self.transfer_state)
+        self.send_or_update_message(self.tx_hash, self.transfer_state)
 
     @transition(field=transfer_state, source='*', target='ERROR')
     def state_transfer_error(self):
-        send_or_update_message(self.tx_hash, self.transfer_state)
+        self.send_or_update_message(self.tx_hash, self.transfer_state)
         print('Transfer not completed, reverting payment', flush=True)
 
     @transition(field=transfer_state, source='*', target='RETURNED')
     def state_transfer_returned(self):
-        send_or_update_message(self.tx_hash, self.transfer_state)
+        self.send_or_update_message(self.tx_hash, self.transfer_state)
 
     @transition(field=transfer_state, source='*', target='IN_QUEUE')
     def state_transfer_in_queue(self):
-        send_or_update_message(self.tx_hash, self.transfer_state)
+        self.send_or_update_message(self.tx_hash, self.transfer_state)
 
     @transition(field=transfer_state, source=['IN_QUEUE',], target='IN_PROCESS')
     def state_transfer_in_process(self):
-        send_or_update_message(self.tx_hash, self.transfer_state)
+        self.send_or_update_message(self.tx_hash, self.transfer_state)
 
     @transition(field=collection_state, source=['NOT_COLLECTED', 'ERROR'], target='COLLECTED')
     def state_collect_duc(self):
