+++ requirements.txt
@@ -94,7 +94,7 @@ pymongo==3.11.4
 pyparsing==2.4.6
 pyrsistent==0.15.7
 pysha3==1.0.2
-pyTelegramBotAPI==3.7.9
+pyTelegramBotAPI==4.1.1
 pytest==5.3.5
 python-bitcoinrpc==1.0
 python-dateutil==2.8.1
